Split tests over classes.
